backend/requirements.txt
alembic==1.14.0
amqp==5.3.1
aniso8601==9.0.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Cors==5.0.0
Flask-DebugToolbar==0.16.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.5.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.2
marshmallow-sqlalchemy==1.1.0
packaging==24.2
passlib==1.7.4
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
regex==2024.11.6
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
